nar/list*: move to nar/ls, add doc
This isn't really a part of the .nar format itself, but mostly a
directory/index for it, shipped in a separate file.
