+++ README.md
@@ -6,4 +6,5 @@ This repository holds a bunch of experiments written in Go.
 
 * nar - a Nix ARchive (NAR) file parser that mimics archive/tar from the
   stdlib
+* nar/ls - a parser for .ls files (providing an index for .nar files)
 * libstore - a narinfo parser
