+++ nar/ls/doc.go
@@ -1,0 +1,7 @@
+// Package ls implements a parser for the .ls file format, which provides an
+// index into .nar files.
+
+// It is provided on cache.nixos.org, and more generally, written when
+// write-nar-listing=1 is passed while copying build results into a binary
+// cache.
+package ls
